Find pip3 when running in a virtualenv.
If autoupgrade is called from a program executed from a virtualenv
like this (where the venv is not activated):

$ venv/bin/myprogram

Without this patch, it would try to call the system pip3, it would
not find the relevant packages installed, and it would crash
(sys.exit(-1), which might not be the best thing to do anyway).

With this patch, it will find the pip3 from the venv, and call it.